1.2.1
- Removed give methods from command classes and added them to Main class. All command classes have a Main instance now.
