+++ src/plugin.yml
@@ -1,7 +1,7 @@
 main: tech.secretgarden.stash.Main
 name: Stash
 author: ChuckLingBao
-version: 1.2.0
+version: 1.2.1
 api-version: 1.17
 commands:
   stash:
