+++ src/tech/secretgarden/stash/StashCommand.java
@@ -11,10 +11,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.Map;
-import java.util.Set;
 
 public class StashCommand implements CommandExecutor {
 
@@ -26,17 +23,12 @@ public class StashCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String label, String args[]) {
 
-        Set<String> configKeys = plugin.getConfig().getKeys(false);
-
         if (sender instanceof Player) {
             Player player = (Player) sender;
             Inventory senderStash = MapConversion.map.get(player.getUniqueId().toString());
             //opens your own stash
             if (args.length == 0) {
                 player.openInventory(senderStash);
-
-                //*********************ADD COMMAND TO OPEN OTHER'S STASH HERE ***************************
-
 
             } else if (args.length == 1 && player.hasPermission("stash.a")) {
                 String p = args[0];
@@ -66,9 +58,9 @@ public class StashCommand implements CommandExecutor {
                             Inventory stashInv = (Inventory) entry.getValue();
                             stashInv.addItem(item);
                         }
-                        configAddAll(configKeys, itemName, integer);
+                        plugin.configAddAll(itemName, integer);
                     } else if (args.length == 4) {
-                        parseIntegersAll(args, item, configKeys, player);
+                        plugin.parseIntegersAll(args, item, player);
                     }
                 /*
                     BELOW IS FOR GIVING ITEMS TO A SINGLE PLAYER
@@ -80,9 +72,9 @@ public class StashCommand implements CommandExecutor {
                         if (args.length == 3) {
                             Integer integer = 1;
                             singleStash.addItem(item);
-                            configAdd(args, itemName, integer);
+                            plugin.configAdd(args, itemName, integer);
                         } else if (args.length == 4) {
-                            parseIntegersSingle(args, singleStash, item, player);
+                            plugin.parseIntegersSingle(args, singleStash, item, player);
                         }
                     } else {
                         //target == null
@@ -95,7 +87,7 @@ public class StashCommand implements CommandExecutor {
                                 //Now see if offlinePlayer String matches map key.
                                 if (MapConversion.map.containsKey(offlinePlayerId)) {
                                     singleStash.addItem(item);
-                                    configAdd(args, itemName, integer);
+                                    plugin.configAdd(args, itemName, integer);
                                 } else {
                                     player.sendMessage("This player has not logged in before");
                                 }
@@ -104,7 +96,7 @@ public class StashCommand implements CommandExecutor {
                             }
                         } else if (args.length == 4) {
                             if (Bukkit.getOfflinePlayer(args[1]).hasPlayedBefore()) {
-                                parseIntegersSingle(args, singleStash, item, player);
+                                plugin.parseIntegersSingle(args, singleStash, item, player);
                             } else {
                                 player.sendMessage(ChatColor.RED + "This is not a valid player.");
                             }
@@ -116,54 +108,5 @@ public class StashCommand implements CommandExecutor {
             }
         }
         return false;
-    }
-
-    public void configAddAll(Set<String> configKeys, String itemName, Integer integer) {
-        LocalDateTime date = LocalDateTime.now();
-        String dateStr = date.format(DateTimeFormatter.ofPattern("EEEE MMMM dd yyyy hh,mm,ss a"));
-        for (String p : configKeys) {
-            plugin.getConfig().set(p + ".Added Items." + "- " + "x" + integer + " " + itemName, dateStr);
-            plugin.saveConfig();
-        }
-    }
-    public void configAdd(String[] args, String itemName, Integer integer) {
-        LocalDateTime date = LocalDateTime.now();
-        String dateStr = date.format(DateTimeFormatter.ofPattern("EEEE MMMM dd yyyy hh,mm,ss a"));
-        String p = args[1];
-        plugin.getConfig().set(p + ".Added Items." + "- " + "x" + integer + " " + itemName, dateStr);
-        plugin.saveConfig();
-    }
-    public void parseIntegersAll(String[] args, ItemStack item, Set<String> keys, Player player) {
-        try {
-            Integer.parseInt(args[3]);
-            int integer = Integer.parseInt(args[3]);
-            String itemName = item.getItemMeta().getDisplayName();
-            if (integer <= item.getMaxStackSize()) {
-                for (Map.Entry entry : MapConversion.map.entrySet()) {
-                    Inventory stashInv = (Inventory) entry.getValue();
-                    for (int i = 0; i < integer; i++) {
-                        stashInv.addItem(item);
-                    }
-                }
-                configAddAll(keys, itemName, integer);
-            }
-        } catch (NumberFormatException nfe) {
-            player.sendMessage("This argument must be an integer");
-        }
-    }
-    public void parseIntegersSingle(String[] args, Inventory singleStash, ItemStack item, Player player) {
-        try {
-            Integer.parseInt(args[3]);
-            String itemName = item.getItemMeta().getDisplayName();
-            int integer = Integer.parseInt(args[3]);
-            if (integer <= item.getMaxStackSize()) {
-                for (int i = 0; i < integer; i++) {
-                    singleStash.addItem(item);
-                }
-                configAdd(args, itemName, integer);
-            }
-        } catch (NumberFormatException nfe) {
-            player.sendMessage("This argument must be an integer");
-        }
     }
 }
