+++ src/tech/secretgarden/stash/StashSfCommand.java
@@ -11,10 +11,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.Map;
-import java.util.Set;
 
 public class StashSfCommand implements CommandExecutor {
     private final Main plugin;
@@ -22,15 +19,8 @@ public class StashSfCommand implements CommandExecutor {
         this.plugin = instance;
     }
 
-    private StashCommand stashCommand;
-    public void setStashCommand(StashCommand stashCommand) {
-        this.stashCommand = stashCommand;
-    }
-
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
-
-        Set<String> keys = plugin.getConfig().getKeys(false);
 
         if (sender instanceof Player) {
             Player player = (Player) sender;
@@ -48,12 +38,11 @@ public class StashSfCommand implements CommandExecutor {
                             Inventory stashInv = (Inventory) entry.getValue();
                             stashInv.addItem(item);
                         }
-                        configAddAll(keys, itemName, integer);
+                        plugin.configAddAll(itemName, integer);
                     } else if (args.length == 4) {
                         //checking if int arg is an Integer
-                        parseIntegersAll(args, item, keys, player);
+                        plugin.parseIntegersAll(args, item, player);
                     }
-                    //END OF SF ALL ARGS
                 }
                 /*
                     BELOW IS FOR GIVING ITEMS TO A SINGLE PLAYER
@@ -63,9 +52,9 @@ public class StashSfCommand implements CommandExecutor {
                     if (args.length == 3) {
                         Integer integer = 1;
                         singleStash.addItem(item);
-                        configAdd(args, itemName, integer);
+                        plugin.configAdd(args, itemName, integer);
                     } else if (args.length == 4) {
-                        parseIntegersSingle(args, singleStash, item, player);
+                        plugin.parseIntegersSingle(args, singleStash, item, player);
                     }
                 } else {
                     OfflinePlayer offlinePlayer = Bukkit.getOfflinePlayer(args[1]);
@@ -78,7 +67,7 @@ public class StashSfCommand implements CommandExecutor {
                             //Now see if offlinePlayer String matches map key.
                             if (MapConversion.map.containsKey(offlinePlayerId)) {
                                 singleStash.addItem(item);
-                                configAdd(args, itemName, integer);
+                                plugin.configAdd(args, itemName, integer);
                             } else {
                                 player.sendMessage("This player has not logged in before");
                             }
@@ -87,7 +76,7 @@ public class StashSfCommand implements CommandExecutor {
                         }
                     } else if (args.length == 4) {
                         if (Bukkit.getOfflinePlayer(args[1]).hasPlayedBefore()) {
-                            parseIntegersSingle(args, singleStash, item, player);
+                            plugin.parseIntegersSingle(args, singleStash, item, player);
                         } else {
                             player.sendMessage(ChatColor.RED + "This is not a valid player.");
                         }
@@ -97,101 +86,6 @@ public class StashSfCommand implements CommandExecutor {
                 player.sendMessage(ChatColor.RED + "You do not have permission.");
             }
         }
-
         return false;
     }
-    public void configAddAll(Set<String> configKeys, String itemName, Integer integer) {
-        LocalDateTime date = LocalDateTime.now();
-        String dateStr = date.format(DateTimeFormatter.ofPattern("EEEE MMMM dd yyyy hh,mm,ss a"));
-        for (String p : configKeys) {
-            plugin.getConfig().set(p + ".Added Items." + "- " + "x" + integer + " " + itemName, dateStr);
-            plugin.saveConfig();
-        }
-    }
-    public void configAdd(String[] args, String itemName, Integer integer) {
-        LocalDateTime date = LocalDateTime.now();
-        String dateStr = date.format(DateTimeFormatter.ofPattern("EEEE MMMM dd yyyy hh,mm,ss a"));
-        String p = args[1];
-        plugin.getConfig().set(p + ".Added Items." + "- " + "x" + integer + " " + itemName, dateStr);
-        plugin.saveConfig();
-    }
-    public void parseIntegersAll(String[] args, ItemStack item, Set<String> keys, Player player) {
-        try {
-            Integer.parseInt(args[3]);
-            int integer = Integer.parseInt(args[3]);
-            String itemName = item.getItemMeta().getDisplayName();
-            if (integer <= item.getMaxStackSize()) {
-                for (Map.Entry entry : MapConversion.map.entrySet()) {
-                    Inventory stashInv = (Inventory) entry.getValue();
-                    for (int i = 0; i < integer; i++) {
-                        stashInv.addItem(item);
-                    }
-                }
-                configAddAll(keys, itemName, integer);
-            }
-        } catch (NumberFormatException nfe) {
-            player.sendMessage("This argument must be an integer");
-        }
-    }
-    public void parseIntegersSingle(String[] args, Inventory singleStash, ItemStack item, Player player) {
-        try {
-            Integer.parseInt(args[3]);
-            String itemName = item.getItemMeta().getDisplayName();
-            int integer = Integer.parseInt(args[3]);
-            if (integer <= item.getMaxStackSize()) {
-                for (int i = 0; i < integer; i++) {
-                    singleStash.addItem(item);
-                }
-                configAdd(args, itemName, integer);
-            }
-        } catch (NumberFormatException nfe) {
-            player.sendMessage("This argument must be an integer");
-        }
-    }
-    /*
-    private void configAddAll(Set<String> keys, ItemStack item, Integer integer, String dateStr) {
-        for (String p : keys) {
-            plugin.getConfig().createSection(p + ".Added Items." + "- " + item);
-            plugin.getConfig().set(p + ".Added Items." + "- " + "x" + integer+ " " + item, dateStr);
-            plugin.saveConfig();
-        }
-    }
-    private void configAdd(String[] args, ItemStack item, Integer integer, String dateStr) {
-        String p = args[1];
-        plugin.getConfig().createSection(p + ".Added Items." + "- " + item);
-        plugin.getConfig().set(p + ".Added Items." + "- " + "x" + integer+ " " + item, dateStr);
-        plugin.saveConfig();
-    }
-    private void parseIntegersAll(String[] args, ItemStack item, Set<String> keys, String dateStr, Player player) {
-        try {
-            Integer.parseInt(args[3]);
-            int integer = Integer.parseInt(args[3]);
-            if (integer <= item.getMaxStackSize()) {
-                for (Map.Entry entry : MapConversion.map.entrySet()) {
-                    Inventory stashInv = (Inventory) entry.getValue();
-                    for (int i = 0; i < integer; i++) {
-                        stashInv.addItem(item);
-                    }
-                }
-                configAddAll(keys, item, integer, dateStr);
-            }
-        } catch (NumberFormatException nfe) {
-            player.sendMessage("This argument must be an integer");
-        }
-    }
-    private void parseIntegersSingle(String[] args, Inventory singleStash, ItemStack item, String dateStr, Player player) {
-        try {
-            Integer.parseInt(args[3]);
-            int integer = Integer.parseInt(args[3]);
-            if (integer <= item.getMaxStackSize()) {
-                for (int i = 0; i < integer; i++) {
-                    singleStash.addItem(item);
-                }
-                configAdd(args, item, integer, dateStr);
-            }
-        } catch (NumberFormatException nfe) {
-            player.sendMessage("This argument must be an integer");
-        }
-    }
-     */
 }
